make civilian agent a separate module
